requirements.txt
Flask==2.3.2
streamlit==1.24.0
opencv-python-headless==4.5.5.62
face-recognition==1.3.0
gspread==5.6.0
google-auth==2.21.0
pandas==1.5.3
requests==2.31.0
Pillow==9.3.0
numpy==1.23.4
